requirements.txt
ecdsa
PyOpenSSL
django-bootstrap3
django-mptt
django-notification
django-session-security
coverage
django-suit
South
django-userena
django-guardian
django-mailgun